cli: print conflicted paths whenever the working copy is changed
This is disabled when the global `--quiet` flag is used.
