+++ CHANGELOG.md
@@ -68,6 +68,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * new function `working_copies()` for revsets to show the working copy commits of all workspaces.
 
+* The list of conflicted paths is printed whenever the working copy changes.
+  This can be disabled with the `--quiet` option.
+
 ### Fixed bugs
 
 ## [0.15.1] - 2024-03-06
